Change constructor renaming to support ghc 7.10
Template Haskell now seems to pass through type operator
names (e.g. :+:) which need to be escaped somehow to become part of a
valid type identifier.

The method of escaping symbols is a little bodgy.

There is still the problem of diffing Integers. This is probably caused
by the changing of internal representation of Integer:
https://ghc.haskell.org/trac/ghc/wiki/Migration/7.10#integer-gmp-1.0.0.0